Improve code style consistency

+++ src/command/MeetupRemoveMentionRole.ts
@@ -5,13 +5,13 @@ import {
     APIApplicationCommandOption,
     ApplicationCommandOptionType,
     ChatInputCommandInteraction,
-    MessageFlags
+    MessageFlags,
 } from "discord.js";
-import {removeRole} from "../cache/meetupAllowedMentionsRoles";
-import {db} from "../database/Database";
-import {tCommand} from "../i18n";
-import {postSuccess} from "../util/postEmbeds";
-import {MeetupAddMentionRoleCommand} from "./MeetupAddMentionRole";
+import { removeRole } from "../cache/meetupAllowedMentionsRoles";
+import { db } from "../database/Database";
+import { tCommand } from "../i18n";
+import { postSuccess } from "../util/postEmbeds";
+import { MeetupAddMentionRoleCommand } from "./MeetupAddMentionRole";
 
 export class MeetupRemoveMentionRoleCommand extends MeetupAddMentionRoleCommand {
     name: string = "meetup_remove_mention_role";
@@ -26,8 +26,8 @@ export class MeetupRemoveMentionRoleCommand extends MeetupAddMentionRoleCommand 
                 name: "role",
                 description: tCommand("meetupRemoveMention.option.roleDescription"),
                 type: ApplicationCommandOptionType.Role,
-                required: true
-            }
+                required: true,
+            },
         ];
     }
 
@@ -37,17 +37,19 @@ export class MeetupRemoveMentionRoleCommand extends MeetupAddMentionRoleCommand 
 
         const { role } = this.sanitizedInputs;
 
-        await db.deleteFrom("meetup_allowed_mentions_role")
-            .where("roleID", "=", role.id)
-            .execute();
+        await db.deleteFrom("meetup_allowed_mentions_role").where("roleID", "=", role.id).execute();
 
         //create success embed
-        await postSuccess(interaction, `Die Rolle <@&${role.id}> ist nun nicht mehr in Meetups erwähnbar`);
+        await postSuccess(
+            interaction,
+            `Die Rolle <@&${role.id}> ist nun nicht mehr in Meetups erwähnbar`,
+        );
     }
 
-    protected async checkInList(roleID: string): Promise<void>{
+    protected async checkInList(roleID: string): Promise<void> {
         //check if role is in list
-        const result = await db.selectFrom("meetup_allowed_mentions_role")
+        const result = await db
+            .selectFrom("meetup_allowed_mentions_role")
             .select("roleID")
             .where("roleID", "=", roleID)
             .execute();
@@ -55,7 +57,9 @@ export class MeetupRemoveMentionRoleCommand extends MeetupAddMentionRoleCommand 
         removeRole(roleID);
 
         if (!result.length) {
-            throw new Error(tCommand("meetupRemoveMention.error.roleAlreadyAdded", {roleID: roleID}));
+            throw new Error(
+                tCommand("meetupRemoveMention.error.roleAlreadyAdded", { roleID: roleID }),
+            );
         }
     }
 }
